refactor: deduplicate dispatchDoctorHeal — keep single copy in commands-handlers.ts
